"Noche de debbuging con Fede (todavia no se arreglo)"

+++ MaPPA/kernel/include/estructuras_kernel.h
@@ -6,17 +6,17 @@
 
 typedef struct { // Todavia no se si esto se va a quedar aca! 
     char* ip_memoria;
-    uint8_t puerto_memoria;
+    int puerto_memoria;
     char* ip_fs;
-    uint8_t puerto_fs;
+    int puerto_fs;
     char* ip_cpu;
-    uint8_t puerto_cpu_ds;
-    uint8_t puerto_cpu_it;
+    int puerto_cpu_ds;
+    int puerto_cpu_it;
     char* alg_planificacion;
-    uint8_t quantum;
+    int quantum;
     t_list* recursos;
     t_list* inst_recursos;
-    uint8_t grado_multip;
+    int grado_multip;
 
 } t_config_k;
